Formulaire infos club ok

+++ branches/new_dao/gestadh45.dao/SaisonDao.cs
@@ -89,7 +89,7 @@ namespace gestadh45.dao
 			cmd.Parameters.Add(paramAnneeFin);
 
 			try {
-				var result = (int)cmd.ExecuteScalar();
+				var result = (long)cmd.ExecuteScalar();
 				return result > 0;
 			}
 			catch (SQLiteException) {
@@ -108,7 +108,7 @@ namespace gestadh45.dao
 			cmd.Parameters.Add(paramId);
 
 			try {
-				var result = (int)cmd.ExecuteScalar();
+				var result = (long)cmd.ExecuteScalar();
 				return result > 0;
 			}
 			catch (SQLiteException) {
